mediatek/mt8183: Add DDR driver of tx rx window perbit cal part
BUG=b:80501386
BRANCH=none
TEST=Boots correctly on Kukui, and inits DRAM successfully with related
     patches.

Reviewed-on: https://review.coreboot.org/c/28842
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/mediatek/mt8183/emi.c
@@ -131,7 +131,7 @@ static void set_rank_info_to_conf(const struct sdram_params *params)
 
 static void set_MRR_pinmux_mapping(void)
 {
-	for (u8 chn = 0; chn < CHANNEL_MAX; chn++) {
+	for (size_t chn = 0; chn < CHANNEL_MAX; chn++) {
 		const u8 *map = phy_mapping[chn];
 		write32(&ch[chn].ao.mrr_bit_mux1,
 			(map[0] << 0) | (map[1] << 8) |
